% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{ApplyLars}
\alias{ApplyLars}
\title{Applies L1-regularised Lasso regression on the dataset.}
\usage{
ApplyLars(data, pred_pos = NULL)
}
\arguments{
\item{data}{a data.frame, columns represent the variables and rows represent the temporal data}

\item{pred_pos}{index of the potential predictor variables in the dataset. If the value is NULL, all variables are considered as potential predictors.}
}
\value{
a matrix of Lasso regression coefficients.
}
\description{
Applies L1-regularised Lasso regression on the dataset.
}
