% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/hello.R
\name{ApplyLasso}
\alias{ApplyLasso}
\title{This function applies L1 regularised Lasso regression on the dataset. The regularisation parameter is selected by cross validation.
It uses function from package "lars".}
\usage{
ApplyLasso(data, pred_pos = NULL)

ApplyLasso(data, pred_pos = NULL)
}
\arguments{
\item{data}{a data.frame, columns represent the variables and rows represent the temporal data}

\item{pred_pos}{index of the potential predictor variables in the dataset. If the value is NULL, all variables are considered as potential predictors.}
}
\value{
a matrix of Lasso regression coefficients.

a matrix of Lasso regression coefficients.
}
\description{
This function applies L1 regularised Lasso regression on the dataset. The regularisation parameter is selected by cross validation.
It uses function from package "lars".

Applies L1-regularised Lasso regression on the dataset.
}
