% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CalculatePrecisionAndRecallForMultiple}
\alias{CalculatePrecisionAndRecallForMultiple}
\title{This function is used to compute the results for a simulation study. It computes individual precision, recall, hamming distance and balanced accuracy
of each learned DBN structure and then calculates the mean and sd of all the individual results.}
\usage{
CalculatePrecisionAndRecallForMultiple(DBNList, datasets)
}
\arguments{
\item{DBNList}{a list of learned DBN structures}

\item{datasets}{a list of datasets simulated using the SimulateData() function. It holds the true DBN network structure (Adjaceny matrix)}
}
\value{
a list of different parameters:"precision_mean","precision_sd","recall_mean", "recall_sd",
"hamming_mean", "hamming_sd", "TN_mean", "TN_sd","FN_mean" , "FN_sd", "FP_mean", "FP_sd", "TP_mean",
"TP_sd", "Accuracy" , "Accuracy_sd".
}
\description{
This function is used to compute the results for a simulation study. It computes individual precision, recall, hamming distance and balanced accuracy
of each learned DBN structure and then calculates the mean and sd of all the individual results.
}
