% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/hello.R
\name{CalculatePriors}
\alias{CalculatePriors}
\title{This function calculates the LOPC or the LASSO parents of Score_LOPC and Score_LASSO}
\usage{
CalculatePriors(allNodes, tarNode, Allcombinations, score_mat, gamma)

CalculatePriors(allNodes, tarNode, Allcombinations, score_mat, gamma)
}
\description{
This function calculates the LOPC or the LASSO parents of Score_LOPC and Score_LASSO

This function calculates the LOPC or the LASSO parents of Score_LOPC and Score_LASSO
}
\keyword{internal}
