% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CombineBICandPriors}
\alias{CombineBICandPriors}
\title{Internal function.
For full exhaustive search, sum up the BIC score of each possible parent
combination and the corresponsding Eij and ¬Eij scores for the edges.}
\usage{
CombineBICandPriors(bic_scores, priors)
}
\arguments{
\item{bic_scores}{BIC scores for each target nodes for each possible parent combination}

\item{all_target_weights}{the E and ¬E calculated for each possible parents combination}

\item{score}{This refers to the type of score to be calculated. The default is "Score_LOPC".
Other possible values are Score_LASSO and BIC.}

\item{gamma}{value of the gamma constant added to the final score, to move the very small edge score values away from zero}

\item{all_parent_combinations}{List of all the combination of parent nodes, with maximum number of parents limited to 3}
}
\description{
Internal function.
For full exhaustive search, sum up the BIC score of each possible parent
combination and the corresponsding Eij and ¬Eij scores for the edges.
}
\keyword{internal}
