% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{CombineBICandWeight}
\alias{CombineBICandWeight}
\title{Internal function.
For full exhaustive search, sum up the BIC score of each possible parent
combination and the corresponsding Eij and ¬Eij scores for the edges.}
\usage{
CombineBICandWeight(
  all_targets_weights,
  bic_scores,
  gamma = 0,
  score = "Score_LOPC",
  all_parent_combinations
)
}
\arguments{
\item{bic_scores}{BIC scores for each target nodes for each possible parent combination}

\item{gamma}{value of the gamma constant added to the final score, to move the very small edge score values away from zero}

\item{score}{This refers to the type of score to be calculated. The default is "Score_LOPC".
Other possible values are Score_LASSO and BIC.}

\item{all_parent_combinations}{List of all the combination of parent nodes, with maximum number of parents limited to 3}

\item{all_target_weights}{the E and ¬E calculated for each possible parents combination}
}
\description{
Internal function.
For full exhaustive search, sum up the BIC score of each possible parent
combination and the corresponsding Eij and ¬Eij scores for the edges.
}
\keyword{internal}
