% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ConstructBNUsingMaxScoringParents}
\alias{ConstructBNUsingMaxScoringParents}
\title{This function takes in all scores for all parents combinations (columns) of all target nodes (rows)
and returns the maximum scoring network.}
\usage{
ConstructBNUsingMaxScoringParents(
  BICandWeight,
  all_parent_combinations,
  allNodes
)
}
\arguments{
\item{all_parent_combinations}{All possible combinations of parents , upto 3 parents max}

\item{allNodes}{all the target nodes in the network}

\item{BICandWeigh}{Score_LASSO or Score_LOPC for each combination of parents for each target node}
}
\description{
This function takes in all scores for all parents combinations (columns) of all target nodes (rows)
and returns the maximum scoring network.
}
\keyword{internal}
