% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBN4GRN.R
\docType{package}
\name{DBN4GRN}
\alias{DBN4GRN}
\title{DBN4GRN Package
contains functions to learn sparse dynamic Bayesian network (DBN) structures to elucidate Gene Reguatory Networks (GRNs) from time series gene expression data.
This package implements and uses two Bayesian network learning scores: Score_Lasso and Score_LOPC and uses them with exhaustive search
and greedy hill-climbing search to learn the structure of a DBN.}
\author{
Hamda Ajmal \email{h.ajmal1@nuigalway.ie}
}
