% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{GreedySearchParentsWithPriors}
\alias{GreedySearchParentsWithPriors}
\title{This function applied greedy hill climbing search to find the parents of each node in the network from the previous time slice.
The score for all arcs absent in the network is calculated as :}
\usage{
GreedySearchParentsWithPriors(
  data,
  score_mat = NULL,
  maxP = Inf,
  gamma = 0,
  pred_pos = NULL,
  score = "Score_BIC"
)
}
\arguments{
\item{data}{a data.frame where each column represents the variables and rows contain the temporal data}

\item{score_mat}{a matrix of scores of edges to be added to the BIC score. This can be the matrix of lasso regression coefficience
or the p-value of the first order conditional independence tests.}

\item{maxP}{maximum number of allowed parents per node}

\item{gamma}{value of the gamma constant added to the final score, to move the very small edge score values away from zero}

\item{pred_pos}{position of the putative regulators in the dataset. If this value is NULL, all variables are considered as putative regulators.}

\item{score}{This refers to the type of score to be calculated. The default is "Score_LOPC".
Other possible values are Score_LASSO and BIC.}
}
\value{
bnlearn object
}
\description{
This function applied greedy hill climbing search to find the parents of each node in the network from the previous time slice.
The score for all arcs absent in the network is calculated as :
}
