% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/hello.R
\name{GreedySearchWithPriors_OneTarget}
\alias{GreedySearchWithPriors_OneTarget}
\title{Internal function
Applied greedy hill-climbing search to each local target node in the netork
\eqn{E_ij = log(score_mat + gamma)}
 \eqn{¬E_ij =  log(1-score_mat+gamma)}}
\usage{
GreedySearchWithPriors_OneTarget(
  score_mat,
  targetIdx,
  all_parents,
  dataS,
  maxP = maxP,
  gamma,
  score
)

GreedySearchWithPriors_OneTarget(
  score_mat,
  targetIdx,
  all_parents,
  dataS,
  maxP = maxP,
  gamma,
  score
)
}
\value{
set of parents arcs for each local target node

set of parents arcs for each local target node
}
\description{
Internal function
Applied greedy hill-climbing search to each local target node in the netork
\eqn{E_ij = log(score_mat + gamma)}
 \eqn{¬E_ij =  log(1-score_mat+gamma)}

Internal function
Applied greedy hill-climbing search to each local target node in the netork
E_ij = log(score_mat + gamma)
¬E_ij =  log(1-score_mat+gamma
}
\keyword{internal}
