% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{MakeAllPossibleCombination}
\alias{MakeAllPossibleCombination}
\title{Makes all possible combination of sets of parents, maximum number of parents set to 3}
\usage{
MakeAllPossibleCombination(all_parents)
}
\description{
Makes all possible combination of sets of parents, maximum number of parents set to 3
}
\keyword{internal}
