% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/hello.R
\name{PercentageValidatedFromRecovered}
\alias{PercentageValidatedFromRecovered}
\title{This function calculates the number and percentage of true positive arcs.}
\usage{
PercentageValidatedFromRecovered(arc_set, validated_set)

PercentageValidatedFromRecovered(arc_set, validated_set)
}
\arguments{
\item{arc_set}{data.frame with two columns, "from" and "to"}

\item{validated_set}{data.frame with two columns, "from" and "to" consisting of pairs of nodes present in the "true" network.}
}
\value{
Number and percentage of TP arcs present in the learned network

number and percentage of TP arcs present in the learned network
}
\description{
This function calculates the number and percentage of true positive arcs.

Calculates the number and percentage of true positive arcs
}
