% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{ScoreEntirepossibleSetsOfParents}
\alias{ScoreEntirepossibleSetsOfParents}
\title{This methods calculates BIC score for each target node, for all the possible combinations of parents}
\usage{
ScoreEntirepossibleSetsOfParents(dataS, tarNode, all_combinations)
}
\description{
This methods calculates BIC score for each target node, for all the possible combinations of parents
}
\keyword{internal}
