% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ShiftData_delta}
\alias{ShiftData_delta}
\title{Puts the data in a time frame with values of variables in the current time slice and their immediate prior value from the previous time slice.}
\usage{
ShiftData_delta(data)
}
\arguments{
\item{data}{a data.frame, columns represent the variables and rows represent the temporal data}
}
\value{
a data.frame with twice the number of columns.
}
\description{
Puts the data in a time frame with values of variables in the current time slice and their immediate prior value from the previous time slice.
}
\keyword{internal}
