% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{SimulateData}
\alias{SimulateData}
\title{Simulate Data
This function first generates a random network and then simulates the multi variate time series data according to the following  first order Auto-Regressive process,
X(t) = AX(t-1) + B + E(t),
where E(t) follows a zero-centered multivariate gaussian distribution whose variance matrix S is diagonal.
It uses different functions from the G1DBN package.}
\usage{
SimulateData(
  genes,
  timepoints,
  seed = 1,
  prop = 0.05,
  range = c(-0.95, -0.05, 0.05, 0.95),
  errors = c(0.03, 0.08)
)
}
\arguments{
\item{genes}{Number of genes/variables}

\item{timepoints}{Number of samples}

\item{seed}{Random seed,  is a number used to initialize a pseudorandom number generator.}

\item{prop}{The proportion of edges in the entire network}

\item{range}{vector with 4 elements specifying range values for the adjacency matrix generation (minimum negative value, maximum negative value, minimum positive value, maximum positive value)}

\item{errors}{vector of 2 elements specifying min and max value of the uniform distribution from which the error variances are drawn}
}
\value{
A list comprising of dataset generated (data) and the network (RealNet) from which the data is generated.
}
\description{
Simulate Data
This function first generates a random network and then simulates the multi variate time series data according to the following  first order Auto-Regressive process,
X(t) = AX(t-1) + B + E(t),
where E(t) follows a zero-centered multivariate gaussian distribution whose variance matrix S is diagonal.
It uses different functions from the G1DBN package.
}
\examples{
SimulateData(genes = 50, timepoints = 20, seed = 1, prop = 0.05, range = c(-0.95,-0.05,0.05,0.95), errors = c(0.03, 0.08) )
}
