% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yeast_arcs_validated.R
\docType{data}
\name{yeasts_arcs_validated}
\alias{yeasts_arcs_validated}
\title{Regulatory links verified by research and experimentation, downloaded from YEASTRACT database
First column represents the regulator genes and the second column represents the target genes
Each row represents a regulatory relationship validated by experimentation and curated in the YEASTRACT database}
\format{
An object of class data.frame. First column represents the regulator genes and the second column represents the target genes
}
\source{
\href{http://www.yeastract.com/index.php}
}
\usage{
data(yeasts_arcs_validated)
}
\description{
Regulatory links verified by research and experimentation, downloaded from YEASTRACT database
First column represents the regulator genes and the second column represents the target genes
Each row represents a regulatory relationship validated by experimentation and curated in the YEASTRACT database
}
\examples{
data(yeasts_arcs_validated)
}
\references{
Monteiro et al. (2020)  Nucleic Acids Research, 48(D1):D642-D649
(\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC25624/})
}
\keyword{datasets}
