% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/hello.R
\name{ExhaustiveSearchForBestParents}
\alias{ExhaustiveSearchForBestParents}
\title{Exhaustively looks for best combination of parents, up to max 3 parents per node.
Currently limits max number of parents to 3. Will be changed in later uppdates.
Please note that this function currently allows you to limited the maximum number of parents to 3
It does not allow you to define a set of possible parents as in the GreedySearchWithPriors function.}
\usage{
ExhaustiveSearchForBestParents(
  data,
  type = "Score_BIC",
  gamma = NULL,
  score_mat = NULL
)

ExhaustiveSearchForBestParents(
  data,
  type = "Score_BIC",
  gamma = NULL,
  score_mat = NULL
)
}
\arguments{
\item{data}{data.frame where columns represent the variables and rows contain the temporal data}

\item{type}{A scoring objective. Possible values are "Score_BIC", "Score_LASSO" and "Score_LOPC".}

\item{gamma}{a constant parameter to be added to the score_mat, to shift the values away from zero. Required for Score_LASSO and Score_LOPC}

\item{score_mat}{a weight matrix to be added to the BIC score, to calculate "Score_LOPC" and "Score_LASSO"}
}
\description{
Exhaustively looks for best combination of parents, up to max 3 parents per node.
Currently limits max number of parents to 3. Will be changed in later uppdates.
Please note that this function currently allows you to limited the maximum number of parents to 3
It does not allow you to define a set of possible parents as in the GreedySearchWithPriors function.

Exhaustively looks for best combination of parents, up to max 3 parents per node.
Currently limits max number of parents to 3. Will be changed in later uppdates.
}
