% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/hello.R
\name{ImputeData}
\alias{ImputeData}
\title{Applies linear interpolation for missing (NA) values in the data. This function uses the package "ImputeTS"}
\usage{
ImputeData(data)

ImputeData(data)
}
\arguments{
\item{data}{data.frame with missing values}
}
\value{
data.frame with missing values imputed with linear interpolation

data.frame with missing values imputed with linear interpolation
}
\description{
Applies linear interpolation for missing (NA) values in the data. This function uses the package "ImputeTS"

Applies linear interpolation for missing (NA) values in the data
}
