% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{PercentageValidatedFromRecovered}
\alias{PercentageValidatedFromRecovered}
\title{This function calculates the number and percentage of true positive arcs.}
\usage{
PercentageValidatedFromRecovered(arc_set, validated_set)
}
\arguments{
\item{arc_set}{data.frame with two columns, "from" and "to", that represents the arcs of the learned DBN.}

\item{validated_set}{data.frame with two columns, "from" and "to" consisting of pairs of nodes present in the "true" network.}
}
\value{
Number and percentage of TP arcs present in the learned network
}
\description{
This function calculates the number and percentage of true positive arcs.
}
